requirements-gpu.txt
absl-py
astor
backcall
bleach
certifi
cycler
decorator
gast
grpcio
grpcio_tools
h5py==2.7.1
html5lib==0.9999999
jedi
Keras==2.1.6
tensorflow-gpu
kiwisolver
Mako
Markdown==2.6.11
MarkupSafe
matplotlib
numpy
opencv-python
pexpect
pickleshare
Pillow
prompt-toolkit
protobuf
ptyprocess
Pygments
pyparsing
python-dateutil
pytz
PyWavelets
PyYAML
scikit-image
scipy
simplegeneric
six
termcolor
traitlets
tqdm
wcwidth
webencodings
Werkzeug
opencv-python
dlib
